feat: Package renamed to animl2
