+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "simple-animl"
+name = "animl2"
 version = "0.0.0"
 authors = [
   { name="Erik Trygg", email="etrg@novonordisk.com" },
@@ -36,7 +36,7 @@ build-backend = "setuptools.build_meta"
 
 [tool.setuptools.packages.find]
 where = ["src"]
-include = ["simple_animl*"]
+include = ["animl2*"]
 
 [tool.isort]
 profile = "black"
