+++ src/animl2/core/__init__.py
