+++ src/animl2/core/annotations.py
