+++ src/animl2/core/base.py
