+++ src/animl2/models/author.py
