+++ src/animl2/models/category.py
