+++ src/animl2/models/device.py
