+++ src/animl2/models/method.py
