+++ src/animl2/models/series.py
