+++ src/animl2/models/technique.py
