+++ src/animl2/models/template.py
