+++ src/animl2/utils/regex.py
