+++ tests/test_document.py
@@ -1,13 +1,7 @@
 import unittest
 
-from simple_animl.core import XmlModel
-from simple_animl.models.doc import (
-    VERSION,
-    XMLNS,
-    XMLNS_XSI,
-    XSI_SCHEMALOCATION,
-    AnIMLDoc,
-)
+from animl2.core import XmlModel
+from animl2.models.doc import VERSION, XMLNS, XMLNS_XSI, XSI_SCHEMALOCATION, AnIMLDoc
 
 
 class TestDoc(unittest.TestCase):
