+++ tests/test_import.py
@@ -5,14 +5,14 @@ import unittest
 
 class TestImport(unittest.TestCase):
     def test_import_core(self):
-        from simple_animl.core import Field, XmlModel
+        from animl2.core import Field, XmlModel
 
         _ = Field, XmlModel
 
     def test_import_all_models(self):
         """Test importing of all *.py files in the models directory."""
 
-        path = importlib.util.find_spec("simple_animl").submodule_search_locations[0]
+        path = importlib.util.find_spec("animl2").submodule_search_locations[0]
         models_dir = os.path.join(path, "models")
 
         for file_name in os.listdir(models_dir):
@@ -20,7 +20,7 @@ class TestImport(unittest.TestCase):
                 continue
             with self.subTest(file_name=file_name):
                 module_name = file_name[:-3]
-                module_path = f"simple_animl.models.{module_name}"
+                module_path = f"animl2.models.{module_name}"
 
                 try:
                     importlib.import_module(module_path)
