+++ tests/test_sample.py
@@ -1,9 +1,9 @@
 import unittest
 from xml.etree import ElementTree
 
-from simple_animl.core import XmlModel
-from simple_animl.models.doc import AnIMLDoc
-from simple_animl.models.sample import Sample, SampleSet
+from animl2.core import XmlModel
+from animl2.models.doc import AnIMLDoc
+from animl2.models.sample import Sample, SampleSet
 
 
 class TestSample(unittest.TestCase):
